+++ frontend/src/App.jsx
@@ -1,24 +1,8 @@
-import Banner from "./pages/Banner"
-import Form from "./pages/Form"
-import Navbar from "./pages/Navbar"
-import FAQS from "./pages/FAQS"
-import Table from "./pages/FlightPlans"
-import Footer from "./pages/Footer"
-import OurPlaces from "./pages/OurPlaces/OurPlaces"
+import { Home } from "./pages";
 
 
 function App() {
-  return (
-    <>
-      <Navbar />
-      <Banner />
-      <Form />
-      <OurPlaces />
-      <FAQS />
-      <Table />
-      <Footer />
-    </>
-  )
+  return <Home />
 }
 
 export default App
