+++ frontend/src/components/Banner.jsx
