+++ frontend/src/components/FAQS.jsx
