+++ frontend/src/components/index.jsx
@@ -1,0 +1,17 @@
+import Banner from "./Banner"
+import Form from "./Form"
+import Navbar from "./Navbar"
+import FAQS from "./FAQS"
+import Table from "./FlightPlans"
+import Footer from "./Footer"
+import OurPlaces from "./OurPlaces"
+
+export {
+    Banner,
+    Form,
+    Navbar,
+    FAQS,
+    Table,
+    Footer,
+    OurPlaces
+}
