+++ frontend/src/pages/Home.jsx
@@ -1,0 +1,18 @@
+import { Navbar, Banner, Form, OurPlaces, FAQS, Table, Footer } from "./../components";
+
+
+function Home() {
+  return (
+    <>
+      <Navbar />
+      <Banner />
+      <Form />
+      <OurPlaces />
+      <FAQS />
+      <Table />
+      <Footer />
+    </>
+  )
+}
+
+export default Home
