+++ frontend/src/pages/index.jsx
@@ -1,0 +1,5 @@
+import Home from "./Home";
+
+export {
+    Home
+}
