+++ frontend/src/services/index.js
@@ -1,0 +1,1 @@
+// Here consume API
